+++ test/blockchains/contracts/UmbrellaFeeds.test.ts
@@ -32,6 +32,7 @@ describe.skip('Umbrella Feeds debug integration tests', () => {
   });
 
   [
+    // ChainsIds.AVALANCHE,
     // ChainsIds.MULTIVERSX,
     // ChainsIds.MASSA,
     // ChainsIds.LINEA,
@@ -57,7 +58,7 @@ describe.skip('Umbrella Feeds debug integration tests', () => {
 
           case ChainsIds.MASSA:
             // expect(Utilities.isAddressEoa(addr)).false;
-            expect(addr.slice(0, 4)).eq('AS');
+            expect(addr.slice(0, 3)).eq('AS1');
             break;
 
           case ChainsIds.CONCORDIUM:
@@ -71,7 +72,7 @@ describe.skip('Umbrella Feeds debug integration tests', () => {
         }
       }).timeout(5000);
 
-      it(`[${chainId}] #requiredSignatures`, async () => {
+      it.skip(`[${chainId}] #requiredSignatures`, async () => {
         expect(await umbrellaFeeds.requiredSignatures()).eq(2);
       });
 
@@ -82,7 +83,7 @@ describe.skip('Umbrella Feeds debug integration tests', () => {
       }).timeout(10000);
 
       it(`[${chainId}] #getManyPriceDataRaw with non exist key`, async () => {
-        const priceDatas = await umbrellaFeeds.getManyPriceDataRaw(['a']);
+        const priceDatas = await umbrellaFeeds.getManyPriceDataRaw(['abcdefghij']);
         console.log(priceDatas);
         if (!priceDatas) throw Error('undefined priceDatas');
 
@@ -97,8 +98,8 @@ describe.skip('Umbrella Feeds debug integration tests', () => {
       }).timeout(10000);
 
       it(`[${chainId}] #getManyPriceDataRaw with existing keys`, async () => {
-        const priceDatas = await umbrellaFeeds.getManyPriceDataRaw(['GOLD-USD', 'aa']);
-        console.log('result', {chainId, priceDatas});
+        const priceDatas = await umbrellaFeeds.getManyPriceDataRaw(['UMB-USD', 'EGLD-USD']);
+        console.log({priceDatas});
         if (!priceDatas) throw Error('undefined priceDatas');
 
         const priceData = priceDatas[0];
@@ -121,10 +122,10 @@ describe.skip('Umbrella Feeds debug integration tests', () => {
         });
 
         it(`[${chainId}] hash empty data`, async () => {
-          const hash = await hasher.apply(chainId, networkId, target, [], []);
+          const hash = hasher.apply(chainId, networkId, target, [], []);
           const contractHash = await umbrellaFeeds.hashData([], []);
 
-          console.log({hash, contractHash});
+          console.log({chainId, networkId, hash, contractHash});
 
           expect(hash).eq(contractHash);
         });
@@ -146,7 +147,7 @@ describe.skip('Umbrella Feeds debug integration tests', () => {
             umbrellaFeeds.hashData(names, priceDatas),
           ]);
 
-          console.log({hash, contractHash});
+          console.log({chainId, networkId, hash, contractHash});
 
           expect(hash).eq(contractHash);
         });
@@ -180,7 +181,7 @@ describe.skip('Umbrella Feeds debug integration tests', () => {
         }).timeout(10000);
       });
 
-      describe('#signData', async () => {
+      describe.skip('#signData', async () => {
         const hasher = new DeviationHasher();
         let signer: DeviationSignerInterface;
         let networkId: number;